Fix tests
This commit replaces the mockclient with the fake client. The fake
client is a stricter implementation of the actual k8s client. Mainly
because the mockclient was not able to return the current scheme, which
is necessary for the namespace override to work.

Thus I was able to find some bugs and fixed them.

- Some errors were not asserted
- We no explicitly check if the finalizer is there after the reconcile