Add SLA reporting tool
This commit contains a simple SLA reporting tool. It connects to a
Prometheus/Thanos/Mimir instance and queries them to gather the SLA
information.

It then creates an ASCIIdoc template and sends it to VSHN's docgen.
There it will be converted to a PDF and sent to an S3 instance.