Hanlde disk resizes for redis
This will add the ability to handle disk resizes for Redis.

As long as the default plan size is used, it will not do anything. But
if `parameters.size.disk` is set it will trigger a job that will
resize the sts properly and then trigger the helm deployment again.

The actual resize of the PVCs is handled by
https://github.com/vshn/statefulset-resize-controller.