Fix display issue for Redis Backups
There was a display issue caused by the classic Go iteration bug. We
attached the same object to each row. This meant if kubectl tired to
display additional collums (such as the namespace when listing all redis
backups over all namespaces) it showed the same data of the last object
for every row.