Fixing unittest with fakeclient
Since controller-runtime v0.15.0 we need to make the fake client aware
of any objects where we want to interact with the `status` subresource.

Without registering those, the fakeclient will just trow a very
unhelpful `not found` error.

Also adjusted the logic for the maintenace unit test, as it previously
dropped errors silently.