Add Minio maintenance
This will add Minio's maintenance to the AppCat. Minio is a bit special
with their releases, as it's a rolling release.

In order to support generic helm maintenances in the future a new module
was created. This module can handle any helm release and update its
tag according to a given compare strategy.