Migrating to Composition Functions beta
The beta had quite a lot of changes. For one the GRPC communication is
now the way Crossplane communicates with functions.

Crossplane also provides an SDK which does a lot of heavy lifting right
out of the box. However due to the way we handle things the official SDK
is not sufficient. It has thus been wrapped in a `ServiceRuntime`.

Functions now have to be packaged in order to work. Some initial make
targets have been added with this change. They can probably be improved
over the next few weeks/months.

Additional changes:
* Naming and wording has been aligned to match Crossplane's upstream
  wording.
* The runtime is no longer divided between `desired` and `observed` sub-
  objects. (up to debate)
* The runtime mostly just tracks the desired changes and does
  marshalling of the objects. Everything else is off-loaded to the
  official go-sdk.
* Only metadata and status of the composite can be changed! No other
  changes are allowed anymore.