Fix quota namespace not found
This fixes a bug where the quota namespace was not found. It also fixes
a bug that would override the namespace with a wrong name.

There's also now an additional check to see if the name of the namespace
is empty or not. This helps mitigating wrong patches from the PnT
function.