Delay creation of the cluster object
This change will delay the creation of the cluster object in such a way
that its dependencies need to be deployed first.

This will avoid some reconcile loops in provider-kubernetes, as it will
go into a feedback loop and only try to deploy a failed object, while
not reconciling the others.