Fix nested SLI Prober
The current implementation of the SLI prober expects a claim. However
with nested services there are no claims, but onlny a composite.

The composite also behaves a bit differently. While the secret
containing the connection details is essentially the samem it provisions
the secret right away, but empty. It gets then filled up gradually with
secrets as they become available. This breaks the old logic to simply
wait for the `StartupGracePeriod` or until the secret is available.

In contrast, a claim only provisions the secret once everything is ready
and synced.

To mitigate this, the SLI prober will now unconditionally wait for
`StartupGracePeriod` before trying to read a secret or connect to a pg
instance. It also switches from reading the claim secret to the
composite secret.