Adjust terminating quotas for our jobs
APPUiO's definition of a terminating pod is
`.spec.activeDeadlineSeconds >= 0`. It also enforces a
`.spec.activeDeadlineSeconds` for all jobs. So essentially all jobs are
measured against the terminating quota.

The pgrepack job from StackGres has a hardcoded resource request for
CPUs of 1000m which is higher than the default 500m quota for
terminating pods. This causes the pgrepack jobs to be stuck because the
pods can't be scheduled.

With this commit we set the initial quota for terminating pods to 1000m.