Fix mergo bug
By default mergo doesn't allow merging a non empty value with an
empty value. For example if a field is a string "hello" it would not
override it, if the to be merged object contains the same field with "".

This works well until someone wants to override a true boolean with a
false boolean, in which case the merge doesn't happen.

By setting `mergo.WithOverwriteWithEmptyValue`, we can force mergo to
merge empty values into noni empty values.

https://github.com/darccio/mergo/issues/249