Ability to disable billing for selected services
With this commit it's possible to disable the billing for selected
services.

It works by deploying a separate config map containing this information
in the `appcat-control` namespace. This config map is then read by the
service composition function and removes the billing accordingly.

For more detailed information please have a look at the docs in the
commit.