Make the name of the parameter clearer
Instance implies, that we need to pass the name of the instance. But
there are helm charts or operators don't take the instance name as the
services names 1:1.

This can lead to missmatching certificates. So by renaming the param
to serviceName, it should be clearer, that the tls certificates should
match the final service name.