Also log if the probe has a timeout
Currently the logs don't indicate at all that the probes are failing if
there are timeouts. We only export that information.

This makes it hard to find these, as the Sloth metrics don't indicate
the type of errors. We have to look at the raw SLI metrics to find out
that there are timeouts.