Don't override pgSettings with observed values
This commit fixes a bug where we overwrote observed pg settings to the
desired pg settings.

This caused custom settings specified in the claim to never reach the
K8s API in the first place. So the observed state always contained the
default settings that StackGres injects into the object.