Don't prepend service name to generic alerts
Before this commit, the logic would prepend the name of the container
to the alertname. The container name is used internally to determine
how to gather the needed metrics. As the container name may not be same
as the name of the service, this could be confusing.

enterach alert contains all the necessary information about the affected
service in the description. That makes prepending the name redundant.