Fix a bug, where you can't edit an existing VSHNPostgreSQL instance
There is currently a bug in the AddConnectionDetails function that takes
the SGcluster from the `observed` rather then the `desired` state.
Since this function is run *after* the `DeployPostgreSQL` function and
takes the cluster configuration from the `observed` state and at the end
writes it to the `desired` state, it will effectively override every
change done to the `SGCluster` object in the `DeployPostgreSQL`
function.

The function should rather get the `SGCluster` object from the desired
stage to avoid this issue.

Signed-off-by: Nicolas Bigler <nicolas.bigler@vshn.ch>