Fix bucket naming clashes for cloudscale and exoscale
If a nested services needed more than one bucket, the provisioning would
fail. That's because the bucket CRs for cloudscale and exoscale patched
value of the `crossplane.io/composite` label to the `metadata.name` of
the resulting objects. As this will always be the name of the top most
composite, there were name clashes.

This change is implemented in a backward compatible fashion. If there
are already existing CRs, it will simply take their name. If the CRs are
freshly created, the new naming will be used.

Also the naming follows what has been implement for the minio buckets.
It takes the value of the `bucketName` as the name of the CRs. As bucket
names should be unique per provider anyway.