Fix issue where we apply claim fields to the comp
We're operating on the claim during the comp functions. For 99% of cases
this doesn't matter.

However if we accidentally add a field that's only in the claim to the
composite, then Crossplane will not like it.

So we remove the spec from the desired composite when we return the
response. As Crossplane only allows to update the `.metadata` and
`.status` objects.

It might result in an empty spec if there are multiple functions
in the composition. However this was not tested.