Fix pointer leaks by using generics
We re-used the same pointer for every single reconcile of any given
service type. This could lead to leaking settings between the instances.
Resulting in instances enabling features that aren't actually enabled in
the claim belonging to the instance.

This change gets rid of the pointer by leveraging generics and reflect.