Set the correct version for the SGCluster
We need to ensure that the SGCluster always contains the "correct"
version and that we don't override it with just a major version.
The reason is, that Stackgres doesn't allow changing the version number
in the `SGCluster` object. Since stackgres changes the version field
during a minor upgrade, it won't be possible anymore to "revert" this
change to a major version only setting.

Signed-off-by: Nicolas Bigler <nicolas.bigler@vshn.ch>