Fix wrong pgbouncer toggle
Unfortunately we inverted the boolean that enables/disables the
pgbouncer.

Currently if `service.disablePgBouncer: false` is set, it will result in
`disableConnectionPooling: true` in the sgcluster. Which in fact
disables the pgbouncer, while it should be enabled in that case.

As both booleans are negative already, they should not be inverted!