Make connection details split cluster compatible
With this change we ensure that we never reference an instnace namespace
directly in the connection details. Instead we hard code it to the
Crossplane namespace. Then there's a new function in the runtime that
will check each managed resource for connection details, if it's set,
it will automatically create a secret in the instance namespace on the
service cluster via the provider-kubernetes.

There's an annaotation that can be set on the managed resources to
disable this functionality for specific resources.