Use separate observer for root password
If another object is referenced via the connectionDetails for a kube
object, it can happen that the kube object can't be deleted, if the
referenced object is deleted first.

In this specific case, we used to reference the secret that StackGres
creates to expose the root password via the connectionDetails of the
`cluster` kube object. However that secret is managed by StackGres. So it
will be created by StackGres once a sgcluster object is created.

This sgcluster object is managed by the `cluster` kube object, which
leads to the race condition.

It can happen that StackGres is faster with removing the secret than
provider-kubernetes in remove the `cluster` kube object. When that
happens the `cluster` kube object doesn't get deleted, because
provider-kubernetes throws observation warnings. These warnings prevent
the deletion of the actual sgcluster object and the wrapping `cluster`
kube object.