Also check for non empty value when wiating for CD
Even thought the field in a connectionDetail map might already be
populated, it might happen that it still contains an empty value.

This commit adds an additional check to the waiting logic to also
make sure that the value is not empty.

This will fix a racecondition where a pod can start with an empty
value on the very first provisioning.