Make sure the certificate is available quickly
This commit adds logic to make the certificate available more quickly.
It uses the same logic as the scaling for the TLS switch. As long as
the certificate is not available it will reconcile very quickly and set
the resource manually to not ready. Once the cert is there it will
reduce the frequency of the reconciles and mark it ready. This ensures
that the instance doesn't show as ready in the GUI as long as the
cert is not available.