Improve responsiveness with high poll interval
This change will make Crossplane react much quicker to changes in
PostgreSQL that will trigger a restart.

The reason this will take long if the poll interval for Crossplane is
set to a high value is because of the asnc restart method we employ.

If a restart is needed for PostgreSQL then Stackgres will set a
`PendingRestart` condition to true on the sgcluster object. With the
default poll rate of 1 minute it will then take about 1 minute for the
restart to actually trigger.

With a poll rate of 1h it would then take 1 hour for the restart to
trigger, because Crossplane will then only poll the latest state of the
sgcluster object every 60 minutes.

The default poll rate does not apply to Managed Resources that are not
marked as ready. Then it will reconcile much quicker.

So if there's a change between the observed and desired state of the
`profile` object we explicitly set the managed resource of the sgcluster
to unready. This will then trigger the actualy restart even quicker than
before.