Make AppCat realtime compatible
The way we serializse the `Object`s leaves a `creationDate: null` in
the inner manifests.

However this will cause a reconcile loop when the realtime mode is
enabled. While not apparante there might also be a reduction in
reconciles with Crossplane in poll mode.

In either case, the values that get removed by this commit are garbage
and can cause issues down the line.