Ignore Changelog for merge to master
When the PR is merged from develop to master the changelog will also
create an entry for that PR as well.

Which will clutter the changelog with "useless" PRs for the merges to
master.

This commit fixes that by adding an ignore label and instruct the
changelog builder to ignore all PRs with the label.