Fix wrong type for PG RecoveryTimeStamp
The `RecoveryTimeStamp` field should be a pointer to a string instead of
a plain string, as otherwise the field is autopopulated with a "" when
the field is not set, resulting in a non-working sgcluster due to an
invalid timeformat for that field.
Using a pointer resolves this issue as the field will not be present at
all when not set.

Signed-off-by: Nicolas Bigler <nicolas.bigler@vshn.ch>