Check generated secrets by fields
Previously the check only checked if the secret as a whole existed.
However there was a race condition where not all fields got applied
during the first creation. Resulting in missing fields.

This change will now check the secrets field by field. As a side effect
this allows us to add new fields to existing secrets as well.