Change category for the billing CRD
The categories array in the CRD acts as a kind of alias for the CRD.

For example, each Crossplane XRD will have a `category` called `claim`.

When a user issues a `kubectl get claim` it will show all claims in the
namespace, no matter which type they are, they get aggregated.

There's an APIserver virtual resource that's also called `appcat`. This
breaks the `kubectl get appcat` call, because it tries to get the
billing CRs instead of the virtual resources provided by the apiserver.

However, `kubectl get appcats` still works as expected.

By renaming the category to `appcatbilling`, the clash will vanish.
After restarting the apiserver `kubectl get appcat` and
`kubectl get appcats` both work as expected.