Fix `setup-filesyste` initcontainer resources
The `setup-filesystem` initcontianer will get the same amount of
resources as the main `patroni` container. But StackGres will only set
that once.

If the requests for patroni get reduced afterward, the `setup-filesystem`
initcontainer will still keep the original resources.

This can lead to situations where the instance has a lot more requests
than actually specified. If an initcontainer with higher requests is
specified than the main conatiners, k8s will use the higher requests for
the whole pod.