Remove PGSSLCERT and PGSSLKEY from nextcloud env variables
The certificate and key found in the config map for stackgres are the
server certificates for the postgres instance. Nextcloud is using them
as the client certificate. This never broke because envoy didn't
validate the client certificate (for what ever reason). With envoy being
disabled, the server rejected the faulty client certificates, preventing
nextcloud from working. Since we do not have mTLS for our other
services, it's easiest to remove this entirely.