Make sure names are escaped during ready check
K8s doesn't allow any `_` in object names. However `_` is allowed for
user names in database systems. These differences have to be escaped
when applying the objects to the API.

However the newly introduced `IsResourceRead()` function did not account
for that. Which resulted in resources not being found. This in turn
caused Crossplane to remove the supposedly non-existing resources from
the API.

Additionally this commit will set all the objects to be applied with the
policy to only create and update the data. Never to delete it. Deletion
will have to be a manual and deliberate operation from now on. This also
ensures feature parity with the PostgreSQL implementation.